New culled triangles rendering method + switch to sync2 + events (split barriers)
The vertex attributes are no more, which allows for 'fake' indices that
are a combination of triangle index and command index. The result is that
there is a single command (with 1 subcommand) to draw all the culled indices,
which allows for better (easier) compaction and much higher number of geometry to
draw in one batch (still WIP).